+++ packages/stream_video_flutter/lib/src/livestream/livestream_player.dart
@@ -16,6 +16,12 @@ typedef LivestreamBackstageBuilder = Widget Function(
   CallState callState,
 );
 
+typedef LivestreamControlsBuilder = Widget Function(
+  BuildContext context,
+  Call call,
+  CallState callState,
+);
+
 /// Creates a widget that allows a user to view a livestream.
 ///
 /// By default, the widget has call controls and other elements including:
@@ -26,37 +32,55 @@ class LivestreamPlayer extends StatefulWidget {
   ///
   /// * [call] is the livestream call intended to be viewed.
   ///
-  /// * [muted] defines if the call is muted by default.
-  ///
   /// * [showParticipantCount] defines if the call should show participant count.
   ///
   /// * [backButtonBuilder] allows you to build a back/close button for closing the livestream.
+  ///
+  /// * [videoPlaceholderBuilder] allows you to build a video placeholder for the
+  /// video renderer. This is useful when the video is not available or
+  /// disconnected. By default, it uses the [StreamUserAvatar] widget
+  ///
+  /// * [livestreamHostsUnavailableBuilder] allows you to build a custom widget when
+  /// a livestream is connected but no hosts have video enabled.
+  ///
+  /// * [livestreamNotConnectedBuilder] allows you to build a custom widget when
+  /// the livestream is not connected. Provides connection state information.
+  ///
+  /// * [videoRendererBuilder] allows you to build a custom video renderer
   ///
   /// * [allowDiagnostics] displays call diagnostics when the widget is double-tapped.
   const LivestreamPlayer({
     super.key,
     required this.call,
-    this.muted = false,
     this.showParticipantCount = true,
     this.backButtonBuilder,
     this.livestreamEndedBuilder,
     this.livestreamBackstageBuilder,
+    this.livestreamControlsBuilder,
+    this.videoPlaceholderBuilder,
+    this.videoRendererBuilder,
+    this.livestreamHostsUnavailableBuilder,
+    this.livestreamNotConnectedBuilder,
     this.allowDiagnostics = false,
     this.onCallDisconnected,
     this.onRecordingTapped,
     this.onFullscreenTapped,
+    this.startInFullscreenMode = false,
   });
 
   /// The livestream call to display.
   final Call call;
 
-  /// Stores if the call should be muted by default
-  final bool muted;
-
   /// Boolean to display participant count.
   ///
   /// Defaults to true.
   final bool showParticipantCount;
+
+  /// Determines whether the livestream should start in fullscreen mode.
+  /// When true, the video will expand to cover the entire available space.
+  /// When false, the video will be contained within its boundaries.
+  /// Defaults to false.
+  final bool startInFullscreenMode;
 
   /// [WidgetBuilder] used to build an action button on the top left side of
   /// the screen.
@@ -67,6 +91,25 @@ class LivestreamPlayer extends StatefulWidget {
 
   /// The builder used to create a custom widget when the livestream is in backstage mode.
   final LivestreamBackstageBuilder? livestreamBackstageBuilder;
+
+  /// The builder used to create custom controls for the livestream player.
+  /// This allows customization of the control UI elements displayed during the livestream.
+  final LivestreamControlsBuilder? livestreamControlsBuilder;
+
+  /// Builder function used to build a video placeholder.
+  final VideoPlaceholderBuilder? videoPlaceholderBuilder;
+
+  /// Builder function used to build a video renderer.
+  final VideoRendererBuilder? videoRendererBuilder;
+
+  /// Builder function used to create a custom widget when a livestream is connected
+  /// but no hosts have video enabled.
+  final LivestreamHostsUnavailableBuilder? livestreamHostsUnavailableBuilder;
+
+  /// Builder function used to create a custom widget when the livestream is not connected.
+  /// Provides connection state information (isMigrating, isReconnecting) that can be
+  /// used to show appropriate status messages.
+  final LivestreamNotConnectedBuilder? livestreamNotConnectedBuilder;
 
   /// The action to perform when the call is disconnected. By default, it pops the current route.
   final void Function(CallDisconnectedProperties)? onCallDisconnected;
@@ -103,39 +146,20 @@ class _LivestreamPlayerState extends State<LivestreamPlayer>
   /// Stores if the livestream is in cover or contain mode.
   bool _fullscreen = false;
 
-  /// Timer for updating duration.
-  late Timer _durationTimer;
-
-  /// Current duration of call.
-  final ValueNotifier<Duration> _duration =
-      ValueNotifier<Duration>(Duration.zero);
-
   @override
   void initState() {
     super.initState();
     _callStateSubscription = call.state.listen(_setState);
     _callState = call.state.value;
+    _fullscreen = widget.startInFullscreenMode;
 
     _connect();
-
-    final now = DateTime.now();
-    final currentTime = now.millisecondsSinceEpoch;
-    final startedAt = _callState.liveStartedAt?.millisecondsSinceEpoch ??
-        _callState.createdAt?.millisecondsSinceEpoch ??
-        now.millisecondsSinceEpoch;
-    _duration.value = Duration(milliseconds: currentTime - startedAt);
-
-    _durationTimer = Timer.periodic(const Duration(seconds: 1), (timer) {
-      _duration.value = _duration.value + const Duration(seconds: 1);
-    });
   }
 
   @override
   void dispose() {
     _callStateSubscription?.cancel();
     _callStateSubscription = null;
-    _durationTimer.cancel();
-    _duration.dispose();
 
     super.dispose();
   }
@@ -206,33 +230,42 @@ class _LivestreamPlayerState extends State<LivestreamPlayer>
               call: call,
               callState: _callState,
               backButtonBuilder: widget.backButtonBuilder,
+              videoPlaceholderBuilder: widget.videoPlaceholderBuilder,
+              videoRendererBuilder: widget.videoRendererBuilder,
+              livestreamHostsUnavailableBuilder:
+                  widget.livestreamHostsUnavailableBuilder,
+              livestreamNotConnectedBuilder:
+                  widget.livestreamNotConnectedBuilder,
               displayDiagnostics: _isStatsVisible,
               videoFit: _fullscreen ? VideoFit.cover : VideoFit.contain,
             ),
-            Align(
-              alignment: Alignment.bottomCenter,
-              child: ValueListenableBuilder<Duration>(
-                valueListenable: _duration,
-                builder: (context, duration, _) {
-                  return LivestreamInfo(
-                    call: call,
-                    callState: widget.call.state.value,
-                    fullscreen: _fullscreen,
-                    onFullscreenTapped: () {
-                      if (widget.onFullscreenTapped != null) {
-                        widget.onFullscreenTapped?.call();
-                      } else {
-                        setState(() {
-                          _fullscreen = !_fullscreen;
-                        });
-                      }
+            widget.livestreamControlsBuilder?.call(context, call, _callState) ??
+                Align(
+                  alignment: Alignment.bottomCenter,
+                  child: StreamBuilder<Duration>(
+                    stream: call.callDurationStream,
+                    builder: (context, snapshot) {
+                      final duration = snapshot.data ?? Duration.zero;
+
+                      return LivestreamInfo(
+                        call: call,
+                        callState: widget.call.state.value,
+                        fullscreen: _fullscreen,
+                        onFullscreenTapped: () {
+                          if (widget.onFullscreenTapped != null) {
+                            widget.onFullscreenTapped?.call();
+                          } else {
+                            setState(() {
+                              _fullscreen = !_fullscreen;
+                            });
+                          }
+                        },
+                        duration: duration,
+                        showParticipantCount: widget.showParticipantCount,
+                      );
                     },
-                    duration: duration,
-                    showParticipantCount: widget.showParticipantCount,
-                  );
-                },
-              ),
-            ),
+                  ),
+                ),
           ],
         ),
       ),