+++ packages/stream_video_flutter/CHANGELOG.md
@@ -1,3 +1,50 @@
+## Unreleased
+
+✅ Added
+* Added `handleCallInterruptionCallbacks` method to `RtcMediaDeviceNotifier` that provides an option to handle system audio interruption like incoming calls, or other media playing. See the [documentation](https://getstream.io/video/docs/flutter/advanced/handling-system-audio-interruptions/) for details.
+* Improved the Picture-in-Picture (PiP) implementation for video calls
+    * (iOS) Shows participant avatar instead of black screen when video track is disabled.
+    * (iOS) Added overlay with participant name, microphone indicator and connection qualit indicator.
+    * (iOS/Android) Added `sort` in `PictureInPictureConfiguration` that enables customization of PiP participant selection.
+
+🐞 Fixed
+* Fixed the handling of user blocking event to disconnect the blocked user with a proper reason.
+
+## 0.9.5
+
+✅ Added
+* Introduced `extendBody` parameter in `StreamCallContent` that extends the participants view to the bottom edge of the scaffold when enabled.
+* Introduced `allowMultipleActiveCalls` option in `StreamVideoOptions` enabling support for concurrent active calls. See the [documentation](https://getstream.io/video/docs/flutter/advanced/multiple-simultaneous-calls-support/) for details.
+
+🐞 Fixed
+* Enhanced permission update handling to only process events targeting the current user.
+* Improved audio output device selection to prioritize external audio devices during call connection.
+
+## 0.9.4
+
+🐞 Fixed
+* Resolved an issue where the screen share track would disappear when zoomed in.
+* Fixed screen sharing layout issues on wide screens.
+* Fixed `ScreenShareCallParticipantsContent` with Spotlight view on very wide windows.
+
+✅ Added
+* Introduced new builder parameters to `LivestreamPlayer` for enhanced customization:
+    - `videoRendererBuilder`: Build a custom video renderer.
+    - `videoPlaceholderBuilder`: Build a placeholder for the video renderer.
+    - `livestreamHostsUnavailableBuilder`: Customize the UI when a livestream is connected but no hosts have video enabled.
+    - `livestreamNotConnectedBuilder`: Customize the UI when the livestream is not connected.
+
+## 0.9.3
+
+✅ Added
+* Introduced `callDurationStream` to the `Call` class. A `Stream<Duration>` that emits the current call duration.
+* Added `createdByUser` property to the `CallState` for better tracking of the initiating user.
+* Added `livestreamControlsBuilder` parameter to `LivestreamPlayer` allowing customisation of controls panel.
+
+🐞 Fixed
+* Resolved an issue where participants’ tracks would briefly disappear ("blink") or change location when many participants are in the call.
+* Fixed a bug where subsequent screen sharing sessions did not end properly when terminated via the browser UI.
+
 ## 0.9.2
 
 🐞 Fixed